refine controller, develop mappers, handle exception

+++ web-layer/src/main/java/com/company/ordermanagementsystem/dto/OrderItemDTO.java
@@ -2,9 +2,10 @@ package com.company.ordermanagementsystem.dto;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record OrderItemDTO(@NotNull UUID productId,
                            int quantity,
-                           double unitPrice) {
+                           BigDecimal unitPrice) {
 }
